+++ deepcave/config.py
@@ -1,10 +1,8 @@
-import os
-# from appdirs import user_cache_dir
+from pathlib import Path
 
-root = os.getcwd()
-cache_dir = os.path.join(root, "cache")  # user_cache_dir("deepcave")
-working_dir = os.path.join(root, "examples", "logs", "DeepCAVE")
-
+root = Path.cwd()
+cache_dir = root / "cache"  # user_cache_dir("deepcave")
+working_dir = root / "examples" / "logs" / "DeepCAVE"
 
 # General information to start services
 CONFIG = {
@@ -19,6 +17,5 @@ META = {
     'run_ids': {},  # {run_name: run_id}
     'groups': {}  # {group_name: [run_name, ...]}
 }
-
 
 __all__ = [CONFIG, META]