+++ deepcave/layouts/not_found.py
@@ -1,11 +1,11 @@
 from dash import html
+from dash.development.base_component import Component
 
-#from deepcave.server.plugins.plugin_manager import PluginManager
 from deepcave.layouts.layout import Layout
 
 
 class NotFoundLayout(Layout):
-    def __call__(self):
+    def __call__(self) -> Component:
         return html.H2('This page does not exists.')