+++ deepcave/layouts/sidebar.py
@@ -1,7 +1,7 @@
-from dash import html
-from dash.dependencies import Input, Output, State
-from dash.exceptions import PreventUpdate
 from dash import dcc
+from dash import html
+from dash.dependencies import Input, Output
+from dash.development.base_component import Component
 
 from deepcave import app, queue
 from deepcave.layouts.layout import Layout
@@ -77,9 +77,11 @@ class SidebarLayout(Layout):
                 if len(jobs) > 0:
                     return [
                         html.Hr(),
-                        html.H6(className='sidebar-heading d-flex justify-content-between align-items-center px-3 mt-4 mb-1 text-muted', children=[
-                            html.Span("Queue Information")
-                        ]),
+                        html.H6(
+                            className='sidebar-heading d-flex justify-content-between align-items-center px-3 mt-4 mb-1 text-muted',
+                            children=[
+                                html.Span("Queue Information")
+                            ]),
                         html.Ul(className='nav flex-column', children=items),
                     ]
 
@@ -87,14 +89,16 @@ class SidebarLayout(Layout):
             except:
                 return
 
-    def __call__(self):
+    def __call__(self) -> list[Component]:
 
         layouts = []
         for category, points in self.nav_points.items():
             layouts += [
-                html.H6(className='sidebar-heading d-flex justify-content-between align-items-center px-3 mt-4 mb-1 text-muted', children=[
-                    html.Span(category)
-                ])
+                html.H6(
+                    className='sidebar-heading d-flex justify-content-between align-items-center px-3 mt-4 mb-1 text-muted',
+                    children=[
+                        html.Span(category)
+                    ])
             ]
 
             point_layouts = []