+++ deepcave/plugins/__init__.py
@@ -1,7 +1,7 @@
-import os
+from pathlib import Path
 
-from deepcave.plugins.plugin import Plugin
 from deepcave.plugins.dynamic_plugin import DynamicPlugin
+from deepcave.plugins.plugin import Plugin
 from deepcave.plugins.static_plugin import StaticPlugin
 from deepcave.utils.importing import auto_import_iter
 
@@ -9,16 +9,13 @@ from deepcave.utils.importing import auto_import_iter
 def read_plugins():
     plugins = {}
 
-    paths = [os.path.join(os.path.abspath(os.curdir),
-                          'deepcave', 'plugins', '*')]
+    paths = [Path(__file__).parent]
 
     for _, obj in auto_import_iter("deepcave.plugins", paths):
 
         if issubclass(obj, Plugin):
             # Plugin itself is a subclass, filter it out
-            if obj.__module__ == DynamicPlugin.__module__ or \
-               obj.__module__ == StaticPlugin.__module__ or \
-               obj.__module__ == Plugin.__module__:
+            if any(obj.__module__ in c.__module__ for c in [DynamicPlugin, StaticPlugin, Plugin]):
                 continue
 
             pos = obj.position()
@@ -36,37 +33,8 @@ def read_plugins():
 
 plugins = read_plugins()
 
-
-def get_plugin_layouts():
-    layouts = {}
-
-    for plugin in plugins:
-        layouts[plugin.id()] = plugin()
-
-    return layouts
-
-
-def get_plugin_names():
-    names = {}
-
-    for plugin in plugins:
-        names[plugin.id()] = plugin.name()
-
-    return names
-
-
-def get_plugin_categories():
-    categories = {}
-
-    for plugin in plugins:
-        categories[plugin.id()] = plugin.category()
-
-    return categories
-
-
-plugin_layouts = get_plugin_layouts()
-plugin_names = get_plugin_names()
-plugin_categories = get_plugin_categories()
-
+plugin_layouts: dict[str, Plugin] = {plugin.id(): plugin() for plugin in plugins}
+plugin_names: dict[str, str] = {plugin.id(): plugin.name() for plugin in plugins}
+plugin_categories: dict[str, str] = {plugin.id(): plugin.category() for plugin in plugins}
 
 __all__ = [plugin_layouts, plugin_names, plugin_categories]