+++ deepcave/utils/hash.py
@@ -1,15 +1,16 @@
 import hashlib
+from pathlib import Path
 
 
-def string_to_hash(string):
+def string_to_hash(string: str) -> str:
     hash_object = hashlib.md5(string.encode())
     return hash_object.hexdigest()
 
 
-def file_to_hash(filename):
-    with open(filename, "rb") as f:
-        hash = hashlib.md5()
+def file_to_hash(filename: Path) -> str:
+    hash = hashlib.md5()
+    with Path(filename).open("rb") as f:
         while chunk := f.read(4082):
             hash.update(chunk)
 
-        return hash.hexdigest()
+    return hash.hexdigest()